son push

+++ ex01/main.cpp
@@ -2,7 +2,7 @@
 
 int main()
 {
-    int N = 4;
+    size_t N = 4;
     Zombie *zh = zombieHorde(N, "omer");
 
     for (size_t i = 0; i < N; i++)
